Separated out component styling into its own file
Each component now has its styling in a separate file.

Also, did some general cleanup, and refactored the `EnterPgn` page and
`PairingSelection` page so that they don't return early
