+++ packages/TorneloScoresheet/App.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { SafeAreaView, useColorScheme } from 'react-native';
 
-import ErrorToast from './src/components/ErrorToast';
+import ErrorToast from './src/components/ErrorToast/ErrorToast';
 import { Colors } from 'react-native/Libraries/NewAppScreen';
-import Toolbar from './src/components/Toolbar';
+import Toolbar from './src/components/Toolbar/Toolbar';
 import { AppModeStateContextProvider } from './src/context/AppModeStateContext';
 import { ErrorContextProvider } from './src/context/ErrorContext';
 import Main from './src/pages/Main';
