+++ packages/TorneloScoresheet/src/components/Sheet/Sheet.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
-import { Modal, StyleSheet, Text, View } from 'react-native';
-import IconButton from './IconButton';
+import { Modal, Text, View } from 'react-native';
+import IconButton from '../IconButton/IconButton';
+import { styles } from './style';
 
 /**
  * A Modal for displaying information to the user.
@@ -33,43 +34,5 @@ const Sheet: React.FC<SheetProps> = ({ visible, content, dismiss, title }) => {
     </Modal>
   );
 };
-
-const styles = StyleSheet.create({
-  backdrop: {
-    backgroundColor: 'rgba(122, 122, 122, 0.3)',
-    height: '100%',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  contentContainer: {
-    backgroundColor: 'white',
-    margin: 200,
-    borderRadius: 20,
-    display: 'flex',
-    flexDirection: 'column',
-  },
-  content: {
-    marginLeft: 20,
-    marginRight: 20,
-    marginBottom: 40,
-    color: 'black',
-  },
-  header: {
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    width: 400,
-  },
-  title: {
-    paddingLeft: 20,
-    fontSize: 30,
-    fontWeight: '700',
-  },
-  exitButton: {
-    padding: 20,
-  },
-});
 
 export default Sheet;
