+++ packages/TorneloScoresheet/src/components/Toolbar/Toolbar.tsx
@@ -1,16 +1,17 @@
 import React, { useState } from 'react';
-import { Image, StatusBar, StyleSheet, Text, View } from 'react-native';
-import { useAppModeState } from '../context/AppModeStateContext';
+import { Image, StatusBar, Text, View } from 'react-native';
+import { useAppModeState } from '../../context/AppModeStateContext';
 import {
   colours,
   ColourType,
   statusBarStyleForColor,
   textColour,
-} from '../style/colour';
-import { primary as primaryFont } from '../style/font';
-import { AppMode } from '../types/AppModeState';
-import IconButton from './IconButton';
-import Sheet from './Sheet';
+} from '../../style/colour';
+import { BLACK_LOGO_IMAGE, WHITE_LOGO_IMAGE } from '../../style/images';
+import { AppMode } from '../../types/AppModeState';
+import IconButton from '../IconButton/IconButton';
+import Sheet from '../Sheet/Sheet';
+import { styles } from './style';
 
 /**
  * The App's toolbar.
@@ -18,9 +19,6 @@ import Sheet from './Sheet';
  * Depending on what mode the app is in, the way the toolbar looks will be different,
  * as well as what actions and buttons are available
  */
-
-const BLACK_LOGO_IMAGE = require('../../assets/images/icon-logo-black-500.png');
-const WHITE_LOGO_IMAGE = require('../../assets/images/icon-logo-white-500.png');
 
 const colourForMode: Record<AppMode, ColourType> = {
   [AppMode.EnterPgn]: colours.secondary,
@@ -74,32 +72,5 @@ const Toolbar: React.FC = () => {
     </>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    paddingLeft: 20,
-    paddingBottom: 20,
-    paddingRight: 20,
-    paddingTop: 34,
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-  },
-  logo: {
-    display: 'flex',
-    flexDirection: 'row',
-  },
-  logoImage: {
-    width: 40,
-    height: 44,
-    marginRight: 24,
-  },
-  logoText: {
-    fontSize: 34,
-    fontFamily: primaryFont,
-    fontWeight: '700',
-  },
-});
 
 export default Toolbar;
