+++ packages/TorneloScoresheet/src/context/AppModeStateContext.tsx
@@ -1,6 +1,6 @@
 import React, { useContext, useState } from 'react';
-import { makeUseEnterPgnState } from '../hooks/appMode/EnterPgnState';
-import { makeUsePairingSelectionState } from '../hooks/appMode/PairingSelectionState';
+import { makeUseEnterPgnState } from '../hooks/appMode/enterPgnState';
+import { makeUsePairingSelectionState } from '../hooks/appMode/pairingSelectionState';
 import { AppModeState, AppMode } from '../types/AppModeState';
 
 // The global state for the app
