+++ packages/TorneloScoresheet/src/hooks/appMode/pairingSelectionState.ts
