+++ packages/TorneloScoresheet/src/style/images.ts
@@ -1,0 +1,2 @@
+export const BLACK_LOGO_IMAGE = require('../../assets/images/icon-logo-black-500.png');
+export const WHITE_LOGO_IMAGE = require('../../assets/images/icon-logo-white-500.png');
